refactor(Audio): move to lib layer
